requirements.txt
MySQL-python
South
django-summernote
django-reversion
html2text
pillow
